# Optimize Duration.From* integer methods