@jskeet 2017-06-06
Parentheses all added. Will dig out the benchmark results and edit top PR comment in a few minutes.